64-bit support added (now iOS 7 support only), removed the Finding Florida collection and add some current exhibitions.
I also changed the opening launch images from black to white...just for
you lauren 😉

+++ HARN Xcode/Classes/Collection View/CollectionViewController.m
@@ -115,7 +115,7 @@
             
             //Creates the detailView that and begins adding everything for it to present to the users since this is templateted (sp?)
             DetailViewController *detailViewController = [segue destinationViewController];
-            detailViewController.title =  [NSString stringWithFormat:@"%u of %u",  indexPath, lengthOfArray ];
+            detailViewController.title =  [NSString stringWithFormat:@"%ld of %ld",  (long)indexPath, (long)lengthOfArray ];
             
             //This will have to be the image stored in the cell
             UIImage *imageToSend = collectionInfo.largeImages[selectedIndexPath.row];
